Repaired overflow error for bulk retrieval data size. Added check for empty sector in WWLevelSet. Added world file handling to WWSector.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@1816 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWindIOS/TAIGA/TAIGA/BulkRetrieverController.m
@@ -124,8 +124,8 @@
 - (void) observeValueForKeyPath:(NSString*)keyPath ofObject:(id)object change:(NSDictionary*)change context:(void*)context
 {
     [progress setProgress:[object progress] animated:YES];
-    [[self detailTextLabel] setText:[[NSString alloc] initWithFormat:@"%d MB,  %d MB remaining",
-                    dataSize, (int) ((1 - [object progress]) * dataSize)]];
+    [[self detailTextLabel] setText:[[NSString alloc] initWithFormat:@"%.0f MB,  %.0f MB remaining",
+                    dataSize, ((1.0 - [object progress]) * dataSize)]];
 }
 
 @end
